+++ darts/models/forecasting/croston.py
@@ -123,7 +123,12 @@ class Croston(FutureCovariatesLocalForecastingModel):
     def supports_multivariate(self) -> bool:
         return False
 
-    def _fit(self, series: TimeSeries, future_covariates: Optional[TimeSeries] = None):
+    def _fit(
+        self,
+        series: TimeSeries,
+        future_covariates: Optional[TimeSeries] = None,
+        **kwargs,
+    ):
         super()._fit(series, future_covariates)
         self._assert_univariate(series)
         series = self.training_series
@@ -135,6 +140,7 @@ class Croston(FutureCovariatesLocalForecastingModel):
                 if future_covariates is not None
                 else None
             ),
+            **kwargs,
         )
 
         return self