+++ darts/tests/models/forecasting/test_local_forecasting_models.py
@@ -647,7 +647,7 @@ class TestLocalForecastingModels:
             (
                 ExponentialSmoothing(),
                 "ExponentialSmoothing(trend=ModelMode.ADDITIVE, damped=False, seasonal=SeasonalityMode.ADDITIVE, "
-                + "seasonal_periods=None, random_state=0, kwargs=None)",
+                + "seasonal_periods=None, random_state=0)",
             ),  # no params changed
             (
                 ARIMA(1, 1, 1),